requirements.txt
kafka-python
faker
pyspark==3.5.0